requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
httpx
python-dotenv
gTTS
twilio
pydantic